viajes

+++ Taxi24RestAPI/Models/ViajeModel.cs
@@ -10,16 +10,16 @@ namespace Taxi24RestAPI.Models
         public int ID { get; set; }
         public int IDConductor { get; set; }
         public int IDPasajero { get; set; }
-        public float UbicacionInicialLatitud { get; set; }
-        public float UbicacionInicialLongitud { get; set; }
-        public float UbicacionFinalLatitud { get; set; }
-        public float UbicacionFinalLongitud { get; set; }
+        public double UbicacionInicialLatitud { get; set; }
+        public double UbicacionInicialLongitud { get; set; }
+        public double UbicacionFinalLatitud { get; set; }
+        public double UbicacionFinalLongitud { get; set; }
         public char EstatusViaje { get; set; }
 
         public ConductorModel Conductor {get;set;}
         public PasajeroModel Pasajero {get;set;}
 
-        public float distanciaRecorrida()
+        public double distanciaRecorrida()
         {
             return 0;
         }
